Moodle 3.1 should always load lib.php
This fixes community issue #2 It was found that the fullscreen button stopped loading after a page was visited, until the Moodle cache was cleared. This has been caused by a combination of two Moodle core changes:

* MDL-50645 - Caches plugins that make navigation changes so that less needs to be loaded
* MDL-49662 - Which removed the legacy way of extending navigation in local plugins

The outcome was that lib.php was only loaded when the cache of plugins that extend the navigation was not present. This change makes the plugin look as though it extends page navigation, so that it will be loaded on pages, this will allow the full screen button to appear.