MDLE-1068 Use jQuery in AMD modules
YUI is deprecated in Moodle so we should move to using jQuery in an AMD module.

The button is now rendered via a template.

I have rationalised the classes used.

Setting preferences via the core_user_update_user_preferences webservice requires that we define the preferences in the user_preference callback.