MDLE-1336 Button does not work well with 3.6 Boost dashboard page
The blocks on the dashboard in Boost have been changed to flex elements rather than block elements, this has resulted in the full screen button pushing the first block in the central column to one side.

It does not seem to be possible to fix this in a way that will be safe and work consistently with many themes so I am removing the display of the button on the dashboard. The page is fully under the control of users, so f they do not like their block layout they can change it themselves.